refactor(InventoryClient): Internalise fast rebalance logic
The InventoryClient is the only consumer of this function; internalising
it makes subsequent refactoring for origin chain repayment simpler.
