+++ src/utils/FillUtils.ts
@@ -1,9 +1,6 @@
 import { HubPoolClient, SpokePoolClient } from "../clients";
-import { hasBinanceRoute } from "../common";
 import { FillStatus, FillWithBlock, SpokePoolClientsByChain, DepositWithBlock, RelayData } from "../interfaces";
-import { Address, CHAIN_IDs, compareAddressesSimple, EMPTY_MESSAGE, TOKEN_SYMBOLS_MAP } from "../utils";
-import { getInventoryEquivalentL1TokenAddress } from "./TokenUtils";
-import { utils as sdkUtils } from "@across-protocol/sdk";
+import { EMPTY_MESSAGE } from "../utils";
 
 export type RelayerUnfilledDeposit = {
   deposit: DepositWithBlock;
@@ -76,40 +73,6 @@ export function depositForcesOriginChainRepayment(
   return (
     deposit.fromLiteChain || !hubPoolClient.l2TokenHasPoolRebalanceRoute(deposit.inputToken, deposit.originChainId)
   );
-}
-
-/**
- * @notice Returns true if after filling this deposit, the repayment can be quickly rebalanced to a different chain.
- * @dev This function can be used by the InventoryClient and Relayer to help determine whether a deposit should
- * be filled or ignored given current inventory allocation levels.
- */
-export function repaymentChainCanBeQuicklyRebalanced(
-  repaymentChainId: number,
-  repaymentToken: Address,
-  hubPoolClient: HubPoolClient
-): boolean {
-  const { chainId: hubChainId } = hubPoolClient;
-  const originChainIsCctpEnabled =
-    sdkUtils.chainIsCCTPEnabled(repaymentChainId) &&
-    compareAddressesSimple(TOKEN_SYMBOLS_MAP.USDC.addresses[repaymentChainId], repaymentToken.toNative());
-  const originChainIsOFTEnabled =
-    sdkUtils.chainIsOFTEnabled(repaymentChainId) &&
-    compareAddressesSimple(TOKEN_SYMBOLS_MAP.USDT.addresses[repaymentChainId], repaymentToken.toNative()) &&
-    repaymentChainId !== CHAIN_IDs.HYPEREVM; // OFT withdrawals from HyperEVM take ~12 hours.
-  // Repayments on Mainnet can be quickly rebalanced via canonical bridges out of L1.
-  if (originChainIsCctpEnabled || originChainIsOFTEnabled || repaymentChainId === hubChainId) {
-    return true;
-  }
-  // If Binance offers a withdrawal route for this (chain, token), inventory repaid on this chain can be
-  // moved off via Binance in place of the canonical slow-withdrawal bridge. This naturally covers BSC
-  // (whose canonical L2 bridge is Binance for every token) as well as per-token Binance routes on
-  // slow-withdrawal chains like Arbitrum, Optimism, and Base.
-  try {
-    const l1Token = getInventoryEquivalentL1TokenAddress(repaymentToken, repaymentChainId, hubChainId);
-    return hasBinanceRoute(repaymentChainId, l1Token);
-  } catch {
-    return false;
-  }
 }
 
 export function getAllUnfilledDeposits(
